+++ utils/te_predict.py
@@ -35,34 +35,36 @@ class Predictor:
         modelo = load_model(modelo)
         predictions = []
         nt_to_token = { 'a':1, 't':2, 'g':3, 'c':4 }
-        with open(in_fasta) as fa:
-            for record in batch_iterator(SimpleFastaParser(fa), batch_size_value):
-                identifiers = []
-                sequences = []
-                for fid, fsq in record:
-                    identifiers.append(fid)
-                    # Tokenize sequences
-                    sequences.append([ nt_to_token[nt] if nt in nt_to_token.keys() else 5 for nt in fsq.lower() ])
-                # Pad sequences
-                padded_seqs = pad_sequences(sequences, padding='post', maxlen = PADVALUE)
-                pred_values = modelo.predict(padded_seqs,
-                                            batch_size = MAX_PRED_BATCH,
-                                            verbose = 1)
-                # Predict labels
-                identifiers = pd.Series(identifiers, name='id')
-                results_df = label_pred_dataframe(identifiers, pred_values, colunas)
-                predictions.append(results_df)
-        predictions = pd.concat(predictions)
-        predictions.to_csv(out_table, index=False, sep='\t')
+        if in_fasta.exists():
+            with open(in_fasta) as fa:
+                for record in batch_iterator(SimpleFastaParser(fa), batch_size_value):
+                    identifiers = []
+                    sequences = []
+                    for fid, fsq in record:
+                        identifiers.append(fid)
+                        # Tokenize sequences
+                        sequences.append([ nt_to_token[nt] if nt in nt_to_token.keys() else 5 for nt in fsq.lower() ])
+                    # Pad sequences
+                    padded_seqs = pad_sequences(sequences, padding='post', maxlen = PADVALUE)
+                    pred_values = modelo.predict(padded_seqs,
+                                                batch_size = batch_size_value if batch_size_value <= MAX_PRED_BATCH else MAX_PRED_BATCH,
+                                                verbose = 1)
+                    # Predict labels
+                    identifiers = pd.Series(identifiers, name='id')
+                    results_df = label_pred_dataframe(identifiers, pred_values, colunas)
+                    predictions.append(results_df)
+            predictions = pd.concat(predictions)
+            predictions.to_csv(out_table, index=False, sep='\t')
 
 
 def get_seq_from_pred(pred_table: str, label: str, reference_fasta: str, out_fasta: str) -> None:
     df = pd.read_table(pred_table)
     label_ids = df.loc[df['prediction']==label]['id'].to_list()
     reference_fasta = Path(reference_fasta)
-    print(f'    Retrieving {label} sequences...')
-    with open(reference_fasta) as fa, open(out_fasta,'w') as sd:
-        for fid, fsq in SimpleFastaParser(fa):
-                if fid in label_ids:
-                    record = f'>{fid}|{label}\n{fsq}\n'
-                    sd.write(record)
+    if label_ids:
+        print(f'    Retrieving {label} sequences...')
+        with open(reference_fasta) as fa, open(out_fasta,'w') as sd:
+            for fid, fsq in SimpleFastaParser(fa):
+                    if fid in label_ids:
+                        record = f'>{fid}|{label}\n{fsq}\n'
+                        sd.write(record)
